refactor command generate test report
